SensorQualityControl/requirements.txt
imutils 
matplotlib 
numpy 
opencv_contrib_python 
Pillow 
easyocr 
ultralytics 
pyserial
keyboard